fix: account for `error` calls in hook_factory
Previously, an `error` call would prevent calls to `repl` because
the stack offset wasn't being reset properly.

This commit fixes that by setting the `offset` to -1 when
an `error` call is encountered, which allows `hook_factory`
instances to call `repl` on the next encountered Lua line.

Also, the `repl` function is now called *only* when a "line"
event is for a Lua function. This means the repl is never used
to inspect a C function (which is pointless).