add hint how to run the debugger_lua.c.lua file
trying to run the debugger_lua.c.lua file from within the embed folder will result in "attempt to index a nil value" errors, which are not obvious, as to why they happen.
by reading through the readme, I've then found out that this file needs to be run from the main folder, which will allow "io.open('debugger.lua')" to not return nil. 

I've added a comment at the top of the file to make that obvious.